backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
google-cloud-vision==3.4.5
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gspread==5.12.0
PyMuPDF==1.23.8
pdfplumber==0.10.3
Pillow==10.1.0
pandas==2.1.3
numpy==1.24.3
python-multipart==0.0.6
requests==2.31.0
logging==0.4.9.6